requirements.txt
djangorestframework
markdown
django-filter
pandas
scipy
numpy
Django
coreapi
django-filter
django-crispy-forms
django-guardian
pygments
django-debug-toolbar
django-extensions
django-estimators
django-background-tasks
psycopg2
scikit-learn